+++ desktop/electron/main.js
@@ -80,7 +80,7 @@ const UploadResp = {
 };
 
 let sleepBlockerId = null;
-const VERSION = '1.4.9';
+const VERSION = '1.5.0';
 const IS_WINDOWS = process.platform === 'win32';
 
 function beginManageOperation(op) {